requirements.txt
pillow
scrapy
